Move server files around
