--- domaindriven/src/DomainDriven/Server/Servant.hs
@@ -1,7 +1,0 @@
-module DomainDriven.Server.Servant (module X)
-where
-
-import DomainDriven.Server.Servant.Api as X
-import DomainDriven.Server.Servant.MapModel as X
-import DomainDriven.Server.Servant.Server as X
-import DomainDriven.Server.Servant.TaggedSumOfApis as X
